Creación del template de listar empleados.

+++ src/main/java/com/gestion/empleados/entities/Empleado.java
@@ -37,7 +37,7 @@ public class Empleado {
 	private String email;
 
 	@NotNull
-	private int telefono;
+	private Long telefono;
 
 	@NotEmpty
 	private String sexo;
